Moves board template and initialisers to a board component
